Modify Login (mvc pattern)
