--- src/main/java/ui/gp/Tab/ViewFactory.java
@@ -1,4 +1,0 @@
-package ui.gp.Tab;
-
-public class ViewFactory {
-}
